requirements.txt
# Core dependencies
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Deep Learning
torch>=2.0.0
transformers>=4.30.0

# Azure Services
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0
azure-core>=1.29.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Optional: For better performance
# torchvision>=0.15.0
# accelerate>=0.20.0

mcp[cli]
fastmcp
loguru
git+https://github.com/illuin-tech/colpali
